refactor: #1 user entity 관련 class 위치 변경
